src/server/services/ai/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
python-multipart>=0.0.9
anthropic>=0.40.0
openai>=1.60.0
